refactor/default_chatter: 模块化拆分 + VLM API 名称修正
